requirements.txt
streamlit
streamlit-option-menu
deta
python-dotenv